app/requirements.txt
wheel # なんか必要らしい
pylint # Linter
yapf # フォーマッタ
django # フレームワーク
django-environ # .envを使えるようにする
pylint-django # django特有のLinter
djangorestframework # REST APIを利用できるようにする
djangorestframework-simplejwt # JWT認証を有効化
selenium # ブラウザ自動化
chromedriver_binary # Chromeのドライバ (Dockerで動かす場合は必要ない)
gunicorn # 本番用WSGI(Web Server Gateway Interface)
mysqlclient # MySQLを使うためのClient